feat : Added module for preserving music albumdata
