+++ music_album_data.rb
@@ -1,0 +1,19 @@
+require 'json'
+require_relative 'music_album'
+
+module MusicAlbumStorage
+    def save_music_albums
+        @json_music_albums = []
+        @music_albums.each do |music_album|
+          @json_music_albums.push({'name' => music_album.name, 'publish_date' => music_album.publish_date, 'on_spotify' => music_album.on_spotify})
+        end
+        File.write('music_album.json', JSON.pretty_generate(@json_music_albums))
+      end
+    
+      def parse_music_albums
+        File.open('music_album.json', 'w') { |f| f.write JSON.pretty_generate([])} unless File.exist?('music_album.json')
+        JSON.parse(File.read('music_album.json')).map do |music_album|
+           @music_albums << MusicAlbum.new(music_album['name'], music_album['publish_date'], music_album['on_spotify'])
+        end
+      end
+end
